Move TrivialKeystoreLoader to loader.rs
