+++ src/testing/tests.rs
@@ -322,7 +322,7 @@ pub mod generic_keystore_tests {
         // which is longer than we want to borrow `keystores` for).
         async fn check_balance<'b, L: 'static + Ledger>(
             keystore: &(
-                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, ()>,
+                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, LoaderMetadata>,
                 Vec<UserPubKey>,
             ),
             expected_coin_balance: u64,
@@ -1248,7 +1248,7 @@ pub mod generic_keystore_tests {
         // for).
         async fn check_balances<'b, L: Ledger + 'static>(
             keystores: &[(
-                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, ()>,
+                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, LoaderMetadata>,
                 Vec<UserPubKey>,
             )],
             balances: &[Vec<U256>],
@@ -1278,7 +1278,7 @@ pub mod generic_keystore_tests {
 
         async fn check_histories<'b, L: Ledger + 'static>(
             keystores: &[(
-                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, ()>,
+                Keystore<'b, impl KeystoreBackend<'b, L> + Sync + 'b, L, LoaderMetadata>,
                 Vec<UserPubKey>,
             )],
             histories: &[Vec<Vec<TransactionHistoryEntry<L>>>],
